fix(vendor-push): various google push fixes

+++ src/main/scala/com/secretapp/backend/data/models/GooglePushCredentials.scala
@@ -1,3 +1,3 @@
 package com.secretapp.backend.data.models
 
-case class GooglePushCredentials(uid: Int, authId: Long, projectId: Int, regId: String)
+case class GooglePushCredentials(uid: Int, authId: Long, projectId: Long, regId: String)
